Moves port socket to protocol crate.
